Fix broken log rotation, using a diff. approach.
Logger.new will only honor rotation arguments if the a filename is passed.
Since I was passing an IO object, this meant rotation was not working.

Instead of changing how the Rails logger is configured, let Rails do its
default behavior, which is more future-proof. To accomplish rotation, use a
new initializer to inspect the actual log file and copy+truncate it if it has
grown beyond 20 MB.